requirements.txt
requests>=2.28.0
PyYAML>=6.0
jinja2>=3.1.0
